+++ data/rule.go
@@ -1,14 +1,20 @@
 package data
 
 import (
+	"fmt"
 	"time"
 )
 
-// Condition defines parameters to look for in a sample. Either SampleType or SampleID
-// (or both) can be set. They can't both be "".
+// Condition defines parameters to look for in a point or a schedule.
 type Condition struct {
-	ID             string
-	Description    string
+	// general parameters
+	ID            string
+	Description   string
+	ConditionType string
+	MinTimeActive float64
+	Active        bool
+
+	// used with point value rules
 	NodeID         string
 	PointType      string
 	PointID        string
@@ -17,8 +23,16 @@ type Condition struct {
 	Operator       string
 	PointValue     float64
 	PointTextValue string
-	MinTimeActive  float64
-	Active         bool
+
+	// used with shedule rules
+	StartTime string
+	EndTime   string
+	Weekdays  []time.Weekday
+}
+
+func (c Condition) String() string {
+	ret := fmt.Sprintf("  COND: %v, V:%v, A:%v\n", c.Description, c.PointValue, c.Active)
+	return ret
 }
 
 // Action defines actions that can be taken if a rule is active.
@@ -34,6 +48,11 @@ type Action struct {
 	PointValueType string
 	PointValue     float64
 	PointTextValue string
+}
+
+func (a Action) String() string {
+	ret := fmt.Sprintf("  ACTION: %v, %v\n", a.Description, a.PointValue)
+	return ret
 }
 
 // RuleConfig contains parts of the rule that a users changes
@@ -58,6 +77,19 @@ type Rule struct {
 	Actions     []Action
 }
 
+func (r Rule) String() string {
+	ret := fmt.Sprintf("Rule: %v\n", r.Description)
+	ret += fmt.Sprintf("  active: %v\n", r.Active)
+	for _, c := range r.Conditions {
+		ret += fmt.Sprintf("%v", c)
+	}
+	for _, a := range r.Actions {
+		ret += fmt.Sprintf("%v", a)
+	}
+
+	return ret
+}
+
 // NodeToRule converts nodes that make up a rule to a node
 func NodeToRule(ruleNode NodeEdge, conditionNodes, actionNodes []NodeEdge) (*Rule, error) {
 	ret := &Rule{}
@@ -79,6 +111,8 @@ func NodeToRule(ruleNode NodeEdge, conditionNodes, actionNodes []NodeEdge) (*Rul
 			switch p.Type {
 			case PointTypeDescription:
 				newCond.Description = p.Text
+			case PointTypeConditionType:
+				newCond.ConditionType = p.Text
 			case PointTypeID:
 				newCond.NodeID = p.Text
 			case PointTypePointType:
@@ -97,6 +131,14 @@ func NodeToRule(ruleNode NodeEdge, conditionNodes, actionNodes []NodeEdge) (*Rul
 				newCond.MinTimeActive = p.Value
 			case PointTypeActive:
 				newCond.Active = FloatToBool(p.Value)
+			case PointTypeStart:
+				newCond.StartTime = p.Text
+			case PointTypeEnd:
+				newCond.EndTime = p.Text
+			case PointTypeWeekday:
+				if p.Value > 0 {
+					newCond.Weekdays = append(newCond.Weekdays, time.Weekday(p.Index))
+				}
 			}
 		}
 		ret.Conditions = append(ret.Conditions, newCond)