+++ db/nats-handler.go
@@ -84,7 +84,49 @@ func (nh *NatsHandler) Connect() (*natsgo.Conn, error) {
 		return nil, fmt.Errorf("Subscribe message error: %w", err)
 	}
 
+	go func() {
+		for {
+			childNodes, err := nh.db.nodeDescendents(nh.db.rootNodeID(), "", false, false)
+			if err != nil {
+				log.Println("Error getting child nodes to run schedule: ", err)
+			} else {
+				for _, c := range childNodes {
+					err := nh.runSchedule(c)
+					if err != nil {
+						log.Println("Error running schedule: ", err)
+					}
+				}
+			}
+			time.Sleep(time.Second * 5)
+		}
+	}()
+
 	return nc, nil
+}
+
+func (nh *NatsHandler) runSchedule(node data.NodeEdge) error {
+	switch node.Type {
+	case data.NodeTypeRule:
+		p := data.Point{Time: time.Now(), Type: data.PointTypeTrigger}
+		err := nh.processRuleNode(node, "", []data.Point{p})
+		if err != nil {
+			return err
+		}
+
+	case data.NodeTypeGroup:
+		childNodes, err := nh.db.nodeDescendents(node.ID, "", false, false)
+		if err != nil {
+			return err
+		}
+		for _, c := range childNodes {
+			err := nh.runSchedule(c)
+			if err != nil {
+				return err
+			}
+		}
+	}
+
+	return nil
 }
 
 func (nh *NatsHandler) setSwUpdateState(id string, state data.SwUpdateState) error {
@@ -499,6 +541,40 @@ func (nh *NatsHandler) reply(subject string, err error) {
 	nh.Nc.Publish(subject, []byte(reply))
 }
 
+func (nh *NatsHandler) processRuleNode(ruleNode data.NodeEdge, sourceNodeID string, points []data.Point) error {
+	conditionNodes, err := nh.db.nodeDescendents(ruleNode.ID, data.NodeTypeCondition,
+		false, false)
+	if err != nil {
+		return err
+	}
+
+	actionNodes, err := nh.db.nodeDescendents(ruleNode.ID, data.NodeTypeAction,
+		false, false)
+	if err != nil {
+		return err
+	}
+
+	rule, err := data.NodeToRule(ruleNode, conditionNodes, actionNodes)
+
+	if err != nil {
+		return err
+	}
+
+	active, changed, err := ruleProcessPoints(nh.Nc, rule, sourceNodeID, points)
+
+	if err != nil {
+		log.Println("Error processing rule point: ", err)
+	}
+
+	if active && changed {
+		err := nh.ruleRunActions(nh.Nc, rule, sourceNodeID)
+		if err != nil {
+			log.Println("Error running rule actions: ", err)
+		}
+	}
+	return nil
+}
+
 func (nh *NatsHandler) processPointsUpstream(currentNodeID, nodeID, nodeDesc string, points data.Points) error {
 	// at this point, the point update has already been written to the DB
 
@@ -509,35 +585,9 @@ func (nh *NatsHandler) processPointsUpstream(currentNodeID, nodeID, nodeDesc str
 	}
 
 	for _, ruleNode := range ruleNodes {
-		conditionNodes, err := nh.db.nodeDescendents(ruleNode.ID, data.NodeTypeCondition,
-			false, false)
+		err := nh.processRuleNode(ruleNode, nodeID, points)
 		if err != nil {
 			return err
-		}
-
-		actionNodes, err := nh.db.nodeDescendents(ruleNode.ID, data.NodeTypeAction,
-			false, false)
-		if err != nil {
-			return err
-		}
-
-		rule, err := data.NodeToRule(ruleNode, conditionNodes, actionNodes)
-
-		if err != nil {
-			return err
-		}
-
-		active, err := ruleProcessPoints(nh.Nc, rule, nodeID, points)
-
-		if err != nil {
-			log.Println("Error processing rule point: ", err)
-		}
-
-		if active {
-			err := nh.ruleRunActions(nh.Nc, rule, nodeID)
-			if err != nil {
-				log.Println("Error running rule actions: ", err)
-			}
 		}
 	}
 
@@ -575,152 +625,5 @@ func (nh *NatsHandler) processPointsUpstream(currentNodeID, nodeID, nodeDesc str
 		}
 	}
 
-	return nil
-}
-
-// ruleProcessPoints runs points through a rules conditions and and updates condition
-// and rule active status. Returns true if point was processed and active is true.
-// Currently, this function only processes the first point that matches -- this should
-// handle all current uses.
-func ruleProcessPoints(nc *natsgo.Conn, r *data.Rule, nodeID string, points data.Points) (bool, error) {
-	for _, p := range points {
-		allActive := true
-		pointProcessed := false
-		for _, c := range r.Conditions {
-			if c.NodeID != "" && c.NodeID != nodeID {
-				continue
-			}
-
-			if c.PointID != "" && c.PointID != p.ID {
-				continue
-			}
-
-			if c.PointType != "" && c.PointType != p.Type {
-				continue
-			}
-
-			if c.PointIndex != -1 && c.PointIndex != int(p.Index) {
-				continue
-			}
-
-			var active bool
-
-			pointProcessed = true
-
-			// conditions match, so check value
-			switch c.PointValueType {
-			case data.PointValueNumber:
-				switch c.Operator {
-				case data.PointValueGreaterThan:
-					active = p.Value > c.PointValue
-				case data.PointValueLessThan:
-					active = p.Value < c.PointValue
-				case data.PointValueEqual:
-					active = p.Value == c.PointValue
-				case data.PointValueNotEqual:
-					active = p.Value != c.PointValue
-				}
-			case data.PointValueText:
-				switch c.Operator {
-				case data.PointValueEqual:
-				case data.PointValueNotEqual:
-				case data.PointValueContains:
-				}
-			case data.PointValueOnOff:
-				condValue := c.PointValue != 0
-				pointValue := p.Value != 0
-				active = condValue == pointValue
-			}
-
-			if !active {
-				allActive = false
-			}
-
-			if active != c.Active {
-				// update condition
-				p := data.Point{
-					Type:  data.PointTypeActive,
-					Time:  time.Now(),
-					Value: data.BoolToFloat(active),
-				}
-
-				err := nats.SendNodePoint(nc, c.ID, p, false)
-				if err != nil {
-					log.Println("Rule error sending point: ", err)
-				}
-			}
-		}
-
-		if pointProcessed {
-			if allActive != r.Active {
-				p := data.Point{
-					Type:  data.PointTypeActive,
-					Time:  time.Now(),
-					Value: data.BoolToFloat(allActive),
-				}
-
-				err := nats.SendNodePoint(nc, r.ID, p, false)
-				if err != nil {
-					log.Println("Rule error sending point: ", err)
-				}
-			}
-		}
-
-		if pointProcessed && allActive {
-			return true, nil
-		}
-	}
-
-	return false, nil
-}
-
-// ruleRunActions runs rule actions
-func (nh *NatsHandler) ruleRunActions(nc *natsgo.Conn, r *data.Rule, triggerNode string) error {
-	for _, a := range r.Actions {
-		switch a.Action {
-		case data.PointValueActionSetValue:
-			if a.NodeID == "" {
-				log.Println("Error, node action nodeID must be set, action id: ", a.ID)
-			}
-			p := data.Point{
-				Time:  time.Now(),
-				Type:  a.PointType,
-				Value: a.PointValue,
-				Text:  a.PointTextValue,
-			}
-			err := nats.SendNodePoint(nc, a.NodeID, p, false)
-			if err != nil {
-				log.Println("Error sending rule action point: ", err)
-			}
-		case data.PointValueActionNotify:
-			// get node that fired the rule
-			triggerNode, err := nh.db.node(triggerNode)
-			if err != nil {
-				return err
-			}
-
-			triggerNodeDesc := triggerNode.Desc()
-
-			n := data.Notification{
-				ID:         uuid.New().String(),
-				SourceNode: a.NodeID,
-				Message:    r.Description + " fired at " + triggerNodeDesc,
-			}
-
-			d, err := n.ToPb()
-
-			if err != nil {
-				return err
-			}
-
-			err = nh.Nc.Publish("node."+r.ID+".not", d)
-
-			if err != nil {
-				return err
-			}
-		default:
-			log.Println("Uknown rule action: ", a.Action)
-		}
-	}
 	return nil
 }